Add root page change dashboard to /[slug]
